+++ pubspec.yaml
@@ -45,6 +45,7 @@ flutter:
 
   assets:
     - map_style.txt
+    - assets/wc.png
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
